Refactor warning pragmas and improve file handling and thread safety
- Update warning pragmas to use `_MSC_VER` instead of `WINDOWS` in `ofxhPropertySuite.h`.
- Add error checking for file operations in `core.cpp`.
- Use lambda function for `TryRender` in `previewautocacher.cpp`.
- Cast `video_threads_.size()` to `size_t` in `rendermanager.cpp`.
- Add `-Wno-deprecated-declarations` to CMake compiler flags.
- Remove unnecessary `friend class RenderManager` and update parameter names in `rendermanager.h`.

+++ app/render/rendermanager.cpp
@@ -147,7 +147,7 @@ size_t RenderManager::GetLightestThreadIndex()
 	size_t best_index = 0;
 	size_t min_size = SIZE_MAX;
 
-	for (size_t i = 0; i < video_threads_.size(); ++i) {
+	for (size_t i = 0; i < static_cast<size_t>(video_threads_.size()); ++i) {
 		size_t size = video_threads_[i]->QueueSize();
 		if (size < min_size) {
 			min_size = size;
